CourseProject/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
idna==3.7
importlib_metadata==7.1.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mysql-client==0.0.1
mysqlclient==2.2.4
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.2
zipp==3.18.1
